Update code with re-written modularized code
Re-wrote a large portion of the scad files to improve modularity and
parametric nature from the config files. I am sure there are a few
things that could still use tweaking but these should generally produce
all the components needed in the BOM properly. Yes I renamed the files
and am a jerk for doing it, sorry.